requirements.txt
fastapi
uvicorn
python-dotenv
httpx
google-auth
google-auth-oauthlib
pyjwt
supabase
python-multipart
passlib
sqlalchemy
bcrypt
secure-smtplib
email-validator
pytest
aiosmtplib
jinja2
pytest-asyncio
google
pydantic
sendgrid
cryptography
requests
itsdangerous
pandas